requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
tqdm>=4.60.0
diffusers>=0.20.0
datasets>=2.0.0
img2dataset>=1.41.0
pyarrow>=12.0.0
huggingface_hub>=0.16.0
pandas>=1.5.0
